# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de Alvaro's Laravel Api",
        "description": "Esta es una api simple hecha con laravel 8.0 con autenticación realizada con Oauth 2.0",
        "contact": {
            "email": "alvarogarcia1010@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Alvaro's Api"
        }
    ],
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registro de usuario",
                "description": "Registro de usuario por correo",
                "operationId": "register",
                "requestBody": {
                    "description": "Campos para realizar registro",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Alvaro García"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "alvarogarcia1010"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1998-09-10"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "7777-7777"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Faltan campos requeridos, campos con formato invalido, o ya existe en la base de datos"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Inicio de sesión",
                "description": "Inicio de sesión con correo electrónico y contraseña",
                "operationId": "login",
                "requestBody": {
                    "description": "Credenciales para iniciar sesión",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error de inicio de sesión"
                    },
                    "422": {
                        "description": "Campos no validos o faltan campos requeridos"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/myuser": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener usuario loggeado",
                "description": "Se obtiene el usuario logeado actualmente con el token",
                "operationId": "getmyuser",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Acceso no autorizado, para acceder a este recurso debe iniciar sesión"
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "description": "Este método cierra la sesión en el servidor del token enviado por el headers",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Acceso no autorizado, para acceder a este recurso debe iniciar sesión"
                    }
                }
            }
        },
        "/api/reset-password": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Recuperar contraseña",
                "description": "Se envia un link de recuperación de contraseña al correo proporcionado",
                "operationId": "resetPassword",
                "requestBody": {
                    "description": "Campos para pedir reestablecer contraseña",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "El correo proporcionado no existe en la base de datos o token invalido"
                    },
                    "422": {
                        "description": "Faltan campos requeridos, campos con formato invalido"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/password/reset": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cambiar contraseña",
                "description": "Se recibe el token enviado por correo y la nueva contraseña para poder reestablecerla",
                "operationId": "changePassword",
                "requestBody": {
                    "description": "Campos para realizar registro",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "token"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "Cadena de caracteres aleatorios"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Faltan campos requeridos, campos con formato invalido, o ya existe en la base de datos"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}